Separated admin & users seed
