+++ server/shared/database/seeds/index.ts
@@ -1,6 +1,8 @@
 import { EntityManager } from '@mikro-orm/postgresql';
+import seedAdmin from './admin';
 import seedUsers from './users';
 
 export default async (em: EntityManager): Promise<void> => {
   await seedUsers(em);
+  await seedAdmin(em);
 };
